refactor(auth): restructure auth components and fix form imports
- Rename SignInForm component to SignIn for consistency
- Fix form imports and structure in signInForm component
- Move ThemeContextProvider to wrap ErrorBoundary for proper context hierarchy
- Improve error handling in auth service by returning meaningful messages
